+++ src/honeypot_auditor/config/ports.py
@@ -33,6 +33,7 @@ PORT_PRESET_IANA: dict[str, int] = {
     "ipp": 631,
     "memcached": 11211,
     "ntp": 123,
+    "ssdp": 1900,
     "tftp": 69,
 }
 
@@ -63,6 +64,7 @@ PORT_PRESET_DOCKER_RESEARCH: dict[str, int] = {
     "ipp": 1631,
     "memcached": 21211,
     "ntp": 1123,
+    "ssdp": 11900,
     "tftp": 1069,
 }